Transaction selection from the Mempool
Added functionality to select transactions from the mempool in an optimal way to maximize fees received by the miner in the given space (4000000 tx in this case). The only constraint being: If a child transaction is present in the selection then all its parent transactions should also be present in that selection.